RavenDB-17977 - move tests
# Conflicts:
#	test/SlowTests/Sharding/Cluster/ReshardingTests.cs
